+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/example/chanel2/flutter2native/IPhoto.java
@@ -1,10 +1,10 @@
-package com.siyehua.spiexample.channel.flutter2native;
+package com.siyehua.example.chanel2.flutter2native;
 import java.util.ArrayList;
  import java.util.HashMap;
  import org.jetbrains.annotations.NotNull;
  import org.jetbrains.annotations.Nullable;
- import com.siyehua.spiexample.channel.ChannelManager.Result;
-public interface TestPreClassName {
+ import com.siyehua.example.chanel2.ChannelManager.Result;
+public interface IPhoto {
  	void aaa();
  
 }