+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/example/chanel3/native2flutter/IPhoto2.java
@@ -1,0 +1,10 @@
+package com.siyehua.example.chanel3.native2flutter;
+import java.util.ArrayList;
+ import java.util.HashMap;
+ import org.jetbrains.annotations.NotNull;
+ import org.jetbrains.annotations.Nullable;
+ import com.siyehua.example.chanel3.ChannelManager.Result;
+public interface IPhoto2 {
+ 	void aaa();
+ 
+}