+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/spiexample1/channel/ChannelManager.java
@@ -1,7 +1,4 @@
-
-
-String javaStr = '''
-package tool;
+package com.siyehua.spiexample1.channel;
 
 import android.os.Handler;
 import android.os.Looper;
@@ -11,7 +8,11 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.UiThread;
 
-import java.lang.reflect.Type;
+import com.siyehua.spiexample1.channel.native2flutter.Fps;
+import com.siyehua.spiexample1.channel.native2flutter.FpsImpl;
+import com.siyehua.spiexample1.channel.native2flutter.Fps2;
+import com.siyehua.spiexample1.channel.native2flutter.Fps2Impl;
+
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -24,7 +25,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodChannel;
 /**
- * custom doc should replace<br>
+ * <br>
  * ChannelManager manager all changer interfaces.<br>
  * add interface impl, use {@link #addChannelImpl(Class, Object)}},<br>
  * get interface impl, use {@link #getChannel(Class)}.<br>
@@ -90,7 +91,7 @@ public class ChannelManager {
         public <T> T parseObject(@Nullable String text, @NonNull Class<T> clazz);
     }
 
-    private static final String channelName = "123456";
+    private static final String channelName = "com.siyehua.spiexample.channel";
     private static final Map<String, Object> channelImplMap = new ConcurrentHashMap<>();
     private static MethodChannel methodChannel;
     private static final Handler handler = new Handler(Looper.getMainLooper());
@@ -308,9 +309,10 @@ public class ChannelManager {
 
 
     static {
-//generated add native2flutter impl in this
+		addChannelImpl(Fps.class, new FpsImpl());
+		addChannelImpl(Fps2.class, new Fps2Impl());
+
     }
 
 }
 
-''';