+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/spiexample1/channel/flutter2native/MyClass.java
@@ -1,9 +1,9 @@
-package com.siyehua.spiexample.channel.flutter2native;
+package com.siyehua.spiexample1.channel.flutter2native;
 import java.util.ArrayList;
  import java.util.HashMap;
  import org.jetbrains.annotations.NotNull;
  import org.jetbrains.annotations.Nullable;
- import com.siyehua.spiexample.channel.ChannelManager.Result;
+ import com.siyehua.spiexample1.channel.ChannelManager.Result;
 public class MyClass {
 	 public  @Nullable InnerClass abc ;
 	 public  @Nullable ArrayList<HashMap<HashMap<ArrayList<String>, Long>, ArrayList<Long>>> aaa ;