+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/spiexample1/channel/flutter2native/TestPreClassName2.java
@@ -1,0 +1,10 @@
+package com.siyehua.spiexample1.channel.flutter2native;
+import java.util.ArrayList;
+ import java.util.HashMap;
+ import org.jetbrains.annotations.NotNull;
+ import org.jetbrains.annotations.Nullable;
+ import com.siyehua.spiexample1.channel.ChannelManager.Result;
+public interface TestPreClassName2 {
+ 	void aaa();
+ 
+}