+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/spiexample1/channel/native2flutter/Fps.java
@@ -1,16 +1,15 @@
-package com.siyehua.spiexample.channel.native2flutter;
+package com.siyehua.spiexample1.channel.native2flutter;
 import java.util.ArrayList;
  import java.util.HashMap;
  import org.jetbrains.annotations.NotNull;
  import org.jetbrains.annotations.Nullable;
- import  com.siyehua.spiexample.channel.flutter2native.*;
- import com.siyehua.spiexample.channel.ChannelManager.Result;
+ import com.siyehua.spiexample1.channel.ChannelManager.Result;
 public interface Fps {
  	void getPageName( @NotNull Long a,  @NotNull Result<String> callback);
 	void getFps( @NotNull Result<Double> callback);
 	void add11( @NotNull Long b);
 	void getPage( @NotNull Result<PageInfo> callback);
-	void getListCustom( @NotNull ArrayList<InnerClass> a,  @NotNull Result<ArrayList<PageInfo>> callback);
+	void getListCustom( @NotNull ArrayList<Long> a,  @NotNull Result<ArrayList<PageInfo>> callback);
 	void getMapCustom( @NotNull Result<HashMap<PageInfo, Long>> callback);
  
 }