+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/spiexample1/channel/native2flutter/Fps2.java
@@ -1,10 +1,9 @@
-package com.siyehua.spiexample.channel.native2flutter;
+package com.siyehua.spiexample1.channel.native2flutter;
 import java.util.ArrayList;
  import java.util.HashMap;
  import org.jetbrains.annotations.NotNull;
  import org.jetbrains.annotations.Nullable;
- import  com.siyehua.spiexample.channel.flutter2native.*;
- import com.siyehua.spiexample.channel.ChannelManager.Result;
+ import com.siyehua.spiexample1.channel.ChannelManager.Result;
 public interface Fps2 {
  	void getPageName( @NotNull HashMap<String, Long> t,  @NotNull String t2,  @NotNull Result<String> callback);
 	void getFps( @NotNull String t,  @NotNull Long a,  @NotNull Result<Double> callback);