+++ example/flutter_app/android/app/src/main/kotlin/com/siyehua/spiexample1/channel/native2flutter/FpsImpl.java
@@ -1,8 +1,7 @@
-package com.siyehua.spiexample.channel.native2flutter;
+package com.siyehua.spiexample1.channel.native2flutter;
 
-import com.siyehua.spiexample.channel.ChannelManager;
-import com.siyehua.spiexample.channel.ChannelManager.Result;
-import com.siyehua.spiexample.channel.flutter2native.*;
+import com.siyehua.spiexample1.channel.ChannelManager;
+import com.siyehua.spiexample1.channel.ChannelManager.Result;
 import java.util.List;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -32,7 +31,7 @@ public class FpsImpl  implements Fps{
 		ChannelManager.invoke(this.getClass().getInterfaces()[0], "getPage", args, callback);
 	}
 	@Override
-	public void getListCustom( @NotNull ArrayList<InnerClass> a,  @NotNull Result<ArrayList<PageInfo>> callback) {
+	public void getListCustom( @NotNull ArrayList<Long> a,  @NotNull Result<ArrayList<PageInfo>> callback) {
 		List args = new ArrayList();
 		args.add(a);
 		ChannelManager.invoke(this.getClass().getInterfaces()[0], "getListCustom", args, callback);