+++ example/flutter_app/ios/Classes/MQQFlutterGen_/MQQFlutterGen_IPhoto2Imp.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+#import "MQQFlutterGen_IPhoto2.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MQQFlutterGen_IPhoto2Imp : NSObject <MQQFlutterGen_IPhoto2>
+
+@end
+
+NS_ASSUME_NONNULL_END