+++ example/flutter_app/ios/Classes/MQQFlutterGen_/MQQFlutterGen_IPhotoImp.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+#import "MQQFlutterGen_IPhoto.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MQQFlutterGen_IPhotoImp : NSObject <MQQFlutterGen_IPhoto>
+
+@end
+
+NS_ASSUME_NONNULL_END