+++ example/flutter_app/ios/Classes/MQQFlutterGen_/MQQFlutterGen_IPhotoImp.m
@@ -1,0 +1,13 @@
+#import "MQQFlutterGen_IPhotoImp.h"
+#import "MQQFlutterGen_ChannelManager.h"
+NS_ASSUME_NONNULL_BEGIN
+
+@implementation MQQFlutterGen_IPhotoImp
+- (void )aaa
+ {
+	NSMutableArray *args = [NSMutableArray array];
+	[[MQQFlutterGen_ChannelManager sharedManager] invokeMethod:@"aaa" args:args completion:nil];
+}
+
+@end
+NS_ASSUME_NONNULL_END