+++ example/flutter_app/ios/Classes/MQQFlutterGen_/flutter2native/MQQFlutterGen_TestPreClassName2.h
@@ -1,0 +1,12 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol MQQFlutterGen_TestPreClassName2 <NSObject>
+@optional
+
+
+- (void )aaa;
+
+@end
+NS_ASSUME_NONNULL_END