+++ example/flutter_app/lib/channel/flutter2native/account.dart
@@ -5,38 +5,49 @@ abstract class IAccount {
 
   Future<String?> getToken();
 
-  void logout(InnerClass abc, List<InnerClass> list, List<List<Map<int, String>>> aaa);
+  void logout(InnerClass abc, List<InnerClass> list,
+      List<List<Map<int, String>>> aaa);
 
   Future<int> getAge();
 
-  Future<InnerClass> getAge2();
+  Future<InnerClass?> getAge2();
 
   Future<List<String>?> getList(InnerClass? abc);
 
-  Future<Map<List<String>?, InnerClass>> getMap();
+  Future<Map<List<String>?, InnerClass?>> getMap();
 
   void setMap(Map<int, bool>? a);
 
   Future<Map<int, bool>> all(List<int>? a, Map<String?, int> b, int? c);
 }
 
-abstract class TestPreClassName {
+abstract class TestPreClassName2 {
   void aaa();
 }
 
 class MyClass {
   /// Note: this method create by SPI, if change Class property or method,
-  /// please remove it. it will be carted by SPI again.
+  /// please remove it. it will be created by SPI again.
   MyClass();
 
   /// Note: this method create by SPI, if change Class property or method,
-  /// please remove it. it will be carted by SPI again.
+  /// please remove it. it will be created by SPI again.
   MyClass.fromJson(Map<String, dynamic> json) {
-    abc = InnerClass.fromJson(json['abc']);
+    if (json['abc'] != null) {
+      abc = InnerClass.fromJson(json['abc']);
+    }
     if (json['aaa'] != null) {
       aaa = [];
       json['aaa'].forEach((v) {
-        aaa!.add((v as Map).map((key, value) => MapEntry((key as Map).map((key, value) => MapEntry((key as List?)?.map((result) => result as String?).toList(), value as int)), (value as List?)?.map((result) => result as int).toList())));
+        aaa!.add((v as Map).map((key, value) =>
+            MapEntry(
+                (key as Map).map((key, value) =>
+                    MapEntry(
+                        (key as List?)
+                            ?.map((result) => result as String?)
+                            .toList(),
+                        value as int)),
+                (value as List?)?.map((result) => result as int).toList())));
       });
     }
     a = json['a'];
@@ -82,47 +93,70 @@ class MyClass {
     if (json['i'] != null) {
       i = {};
       json['i'].forEach((k, v) {
-        i!.update(k as String, (value) => v as int, ifAbsent: () => v as int);
+        i!.update(k as String
+            , (value) => v as int
+            , ifAbsent: () => v as int);
       });
     }
     if (json['i1'] != null) {
       i1 = {};
       json['i1'].forEach((k, v) {
-        i1.update(k as String, (value) => v as int, ifAbsent: () => v as int);
+        i1.update(k as String
+            , (value) => v as int
+            , ifAbsent: () => v as int);
       });
     }
     if (json['i2'] != null) {
       i2 = {};
       json['i2'].forEach((k, v) {
-        i2.update(k as String?, (value) => v as int, ifAbsent: () => v as int);
+        i2.update(k as String?
+            , (value) => v as int
+            , ifAbsent: () => v as int);
       });
     }
     if (json['i3'] != null) {
       i3 = {};
       json['i3'].forEach((k, v) {
-        i3!.update(InnerClass.fromJson(jsonDecode(k.split("___custom___")[1])), (value) => v as int, ifAbsent: () => v as int);
+        i3!.update(InnerClass.fromJson(jsonDecode(k.split("___custom___")[1]))
+            , (value) => v as int
+            , ifAbsent: () => v as int);
       });
     }
     if (json['i4'] != null) {
       i4 = {};
       json['i4'].forEach((k, v) {
-        i4!.update(InnerClass.fromJson(jsonDecode(k.split("___custom___")[1])), (value) => InnerClass.fromJson(jsonDecode(v.split("___custom___")[1])), ifAbsent: () => InnerClass.fromJson(jsonDecode(v.split("___custom___")[1])));
+        i4!.update(InnerClass.fromJson(jsonDecode(k.split("___custom___")[1]))
+            , (value) =>
+            v != null ? InnerClass.fromJson(
+                jsonDecode(v.split("___custom___")[1])) : null
+            , ifAbsent: () =>
+            v != null ? InnerClass.fromJson(
+                jsonDecode(v.split("___custom___")[1])) : null);
       });
     }
     if (json['i5'] != null) {
       i5 = {};
       json['i5'].forEach((k, v) {
-        i5!.update(InnerClass.fromJson(jsonDecode(k.split("___custom___")[1])), (value) => InnerClass.fromJson(jsonDecode(v.split("___custom___")[1])), ifAbsent: () => InnerClass.fromJson(jsonDecode(v.split("___custom___")[1])));
+        i5!.update(k != null ? InnerClass.fromJson(
+            jsonDecode(k.split("___custom___")[1])) : null
+            , (value) =>
+                InnerClass.fromJson(jsonDecode(v.split("___custom___")[1]))
+            , ifAbsent: () =>
+                InnerClass.fromJson(jsonDecode(v.split("___custom___")[1])));
       });
     }
   }
 
   /// Note: this method create by SPI, if change Class property or method,
-  /// please remove it. it will be carted by SPI again.
+  /// please remove it. it will be created by SPI again.
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = Map<String, dynamic>();
     data['abc'] = abc?.toJson();
-    data['aaa'] = this.aaa?.map((v) => v.map((k, v) => MapEntry(k.map((k, v) => MapEntry(k?.map((v) => v).toList(), v)), v?.map((v) => v).toList()))).toList();
+    data['aaa'] = this.aaa?.map((v) =>
+        v.map((k, v) =>
+            MapEntry(
+                k.map((k, v) => MapEntry(k?.map((v) => v).toList(), v)),
+                v?.map((v) => v).toList()))).toList();
     data['a'] = this.a;
     data['b'] = this.b;
     data['c'] = this.c;
@@ -141,6 +175,7 @@ class MyClass {
     data['i5'] = this.i5?.map((k, v) => MapEntry(k?.toJson(), v.toJson()));
     return data;
   }
+
 
   InnerClass? abc;
   List<Map<Map<List<String?>?, int>, List<int>?>>? aaa;