+++ example/flutter_app/lib/channel/native2flutter/other_busniess.dart
@@ -1,0 +1,8 @@
+/// FileConfig
+/// androidSavePath = ./android/app/src/main/kotlin
+/// channelName = com.siyehua.example23.otherChannelName
+/// packageName = com.siyehua.example.chanel3
+
+abstract class IPhoto2 {
+  void aaa();
+}