+++ example/flutter_app/test/widget_test.dart
@@ -16,6 +16,7 @@ void main() async {
     ,
     AndroidPlatformConfig()
       ..savePath = "./android/app/src/main/kotlin" //android save path
+      ..packageName = "com.siyehua.spiexample1.channel"
     ,
     IosPlatformConfig()
       ..iosProjectPrefix = "MQQFlutterGen_" //iOS pre